PO-59: Two charts in same page
