+++ src/components/Sidebar/SidebarItem.tsx
@@ -1,5 +1,3 @@
-// import { NavLink } from 'react-router-dom';
-
 import { NavLink } from "react-router-dom";
 
 interface ItemProps extends React.HTMLAttributes<HTMLAnchorElement> {
