--- src/pages/Reports/WorkOrders/WorkOrdersByLocation/WorkOrdersByLocation.tsx
@@ -1,18 +1,0 @@
-import { ToastContainer } from 'react-toastify';
-import DefaultLayout from '../../../../layout/DefaultLayout';
-import Chart from './Chart';
-
-const WorkOrdersByLocation: React.FC = () => {
-  return (
-    <>
-      <DefaultLayout>
-        <ToastContainer />
-        <div className="mt-4 ">
-          <Chart />
-        </div>
-      </DefaultLayout>
-    </>
-  );
-};
-
-export default WorkOrdersByLocation;
