+++ src/pages/Reports/WorkOrdersByLocation.tsx
@@ -1,17 +1,21 @@
 import { ApexOptions } from 'apexcharts';
 import React from 'react';
 import ReactApexChart from 'react-apexcharts';
-import { useGetWOByLocation } from '../../../../modules/Reports/useGetWOByLocation';
+import { useGetWOByLocation } from '../../modules/Reports/useGetWOByLocation';
 import Datepicker from 'react-tailwindcss-datepicker';
-import { PrincipalButton, SizeButton } from '../../../../components/CustomButtons/PrincipalButton';
+import {
+  PrincipalButton,
+  SizeButton,
+} from '../../components/CustomButtons/PrincipalButton';
 
-const BarChartComponent: React.FC = () => {
-  const { data, date, handleDateChange, today, isLoading, error, reload } = useGetWOByLocation();
+const WorkOrdersByLocation: React.FC = () => {
+  const { data, date, handleDateChange, today, isLoading, error, reload } =
+    useGetWOByLocation();
 
   const options: ApexOptions = {
     chart: {
       type: 'bar',
-      height: 350,
+      height: 300,
       fontFamily: 'Satoshi, sans-serif',
       toolbar: {
         show: false,
@@ -83,7 +87,7 @@ const BarChartComponent: React.FC = () => {
       <div className="mb-4 flex justify-between ">
         <div className="flex-1">
           <h4 className="text-xl font-semibold text-black dark:text-white">
-            Locations Count
+            Work orders by location
           </h4>
         </div>
         <div className="w-100">
@@ -100,24 +104,24 @@ const BarChartComponent: React.FC = () => {
           options={options}
           series={series}
           type="bar"
-          height={350}
+          height={300}
         />
       </div>
       {isLoading && <div className="flex justify-center">Loading...</div>}
-        {error && (
-          <div className="flex items-center justify-center">
-            <p className="mr-3">Error fetching data</p>
-            <PrincipalButton
-              onClick={reload}
-              bgColor="bg-red"
-              size={SizeButton.Small}
-            >
-              Retry
-            </PrincipalButton>
-          </div>
-        )}
+      {error && (
+        <div className="flex items-center justify-center">
+          <p className="mr-3">Error fetching data</p>
+          <PrincipalButton
+            onClick={reload}
+            bgColor="bg-red"
+            size={SizeButton.Small}
+          >
+            Retry
+          </PrincipalButton>
+        </div>
+      )}
     </div>
   );
 };
 
-export default BarChartComponent;
+export default WorkOrdersByLocation;
